+++ src/plugin/ext/preset_load.rs
@@ -6,6 +6,7 @@ use std::ffi::{CStr, CString};
 use std::ptr::NonNull;
 
 use crate::plugin::instance::Plugin;
+use crate::plugin::preset_discovery::LocationValue;
 use crate::util::unsafe_clap_call;
 
 use super::Extension;
@@ -31,14 +32,14 @@ impl<'a> Extension<&'a Plugin<'a>> for PresetLoad<'a> {
 }
 
 impl PresetLoad<'_> {
-    /// Try to load a preet based on a URI and an optional load keys. This information can be
+    /// Try to load a preet based on a location and an optional load key. This information can be
     /// obtained through the preset discovery factory
     /// ([`Library::preset_discovery_factory()`][[crate::plugin::library::Library::preset_discovery_factory()]]).
     /// Load keys are only used for container presets, otherwise they're `None`. The semantics are
     /// similar to loading state.
     #[allow(clippy::wrong_self_convention)]
-    pub fn from_uri(&self, uri: &str, load_key: Option<&str>) -> Result<()> {
-        let uri_cstring = CString::new(uri).context("URI contained internal null bytes")?;
+    pub fn from_location(&self, location: &LocationValue, load_key: Option<&str>) -> Result<()> {
+        let (location_kind, location_ptr) = location.to_raw();
         let load_key_cstring = load_key
             .map(|load_key| {
                 CString::new(load_key).context("Load key contained internal null bytes")
@@ -48,9 +49,10 @@ impl PresetLoad<'_> {
         let preset_load = self.preset_load.as_ptr();
         let plugin = self.plugin.as_ptr();
         let success = unsafe_clap_call! {
-            preset_load=>from_uri(
+            preset_load=>from_location(
                 plugin,
-                uri_cstring.as_ptr(),
+                location_kind,
+                location_ptr,
                 match load_key_cstring.as_ref() {
                     Some(load_key_cstring) => load_key_cstring.as_ptr(),
                     None => std::ptr::null(),
@@ -61,10 +63,11 @@ impl PresetLoad<'_> {
             Ok(())
         } else {
             anyhow::bail!(
-                "'clap_plugin_preset_load::from_uri()' returned false with {}.",
+                "'clap_plugin_preset_load::from_location()' returned false with {}{}.",
+                location,
                 match load_key {
-                    Some(load_key) => format!("URI '{uri}' and load key '{load_key}'"),
-                    None => format!("URI '{uri}'"),
+                    Some(load_key) => format!(" and load key '{load_key}'"),
+                    None => String::new(),
                 },
             );
         }