+++ src/plugin/preset_discovery/provider.rs
@@ -12,7 +12,7 @@ use clap_sys::factory::draft::preset_discovery::clap_preset_discovery_provider;
 
 use super::indexer::{Indexer, IndexerResults};
 use super::metadata_receiver::{MetadataReceiver, PresetFile};
-use super::{Location, LocationUri, PresetDiscoveryFactory, ProviderMetadata};
+use super::{Location, LocationValue, PresetDiscoveryFactory, ProviderMetadata};
 use crate::util::unsafe_clap_call;
 
 /// A preset discovery provider created from a preset discovery factory. The provider is initialized
@@ -125,13 +125,17 @@ impl<'a> Provider<'a> {
     /// and queries metadata for each preset that matches the declared file extensions. The location
     /// must be obtained from [`declared_data()`][Self::declared_data()]. Returns an error if the
     /// plugin triggered any kind of error. The returned map contains a [`PresetFile`] for each of
-    /// the crawled URIs that the plugin declared presets for, which can be either a single preset
-    /// or a container of multiple presets.
-    pub fn crawl_location(&self, location: &Location) -> Result<BTreeMap<String, PresetFile>> {
+    /// the crawled locations that the plugin declared presets for, which can be either a single
+    /// preset or a container of multiple presets.
+    pub fn crawl_location(
+        &self,
+        location: &Location,
+    ) -> Result<BTreeMap<LocationValue, PresetFile>> {
         let mut results = BTreeMap::new();
 
-        let mut crawl_uri = |uri: String| -> Result<()> {
-            let uri_cstring = CString::new(uri.clone()).context("Invalid UTF-8 in URI")?;
+        let location_flags = location.flags;
+        let mut crawl = |location: LocationValue| -> Result<()> {
+            let (location_kind, location_ptr) = location.to_raw();
 
             // There is no 'end of preset' kind of function in the metadata provider, so when
             // the `MetadataReceiver` is dropped it may still need to write a preset file or
@@ -139,13 +143,15 @@ impl<'a> Provider<'a> {
             // theere. This can happen during the drop.
             let mut result = None;
             {
-                let metadata_receiver = MetadataReceiver::new(&mut result, &uri, location);
+                let metadata_receiver =
+                    MetadataReceiver::new(&mut result, &location, location_flags);
 
                 let provider = self.as_ptr();
                 let success = unsafe_clap_call! {
                     provider=>get_metadata(
                         provider,
-                        uri_cstring.as_ptr(),
+                        location_kind,
+                        location_ptr,
                         metadata_receiver.clap_preset_discovery_metadata_receiver_ptr()
                     )
                 };
@@ -153,29 +159,31 @@ impl<'a> Provider<'a> {
                     // TODO: Is the plugin allowed to return false here? If it doesn't have any
                     //       presets it should just not declare any, right?
                     anyhow::bail!(
-                        "The preset provider returned false when fetching metadata for the URI \
-                         '{uri}'.",
+                        "The preset provider returned false when fetching metadata for {location}.",
                     );
                 }
             }
 
             if let Some(preset_file) = result {
                 let preset_file = preset_file.with_context(|| {
-                    format!("Error while fetching fetching metadata for the URI '{uri}'",)
+                    format!("Error while fetching fetching metadata for {location}")
                 })?;
 
-                results.insert(uri, preset_file);
+                results.insert(location, preset_file);
             }
 
             Ok(())
         };
 
-        match &location.uri {
-            LocationUri::File(file_path) => {
+        match &location.value {
+            LocationValue::File(file_path) => {
                 // Single files are queried as is, directories are crawled. If the declared location
                 // does not exist, then that results in a hard error.
-                let metadata = std::fs::metadata(file_path).with_context(|| {
-                    "Could not query metadata for the declared file location '{file_path}'"
+                let file_path_str = file_path
+                    .to_str()
+                    .context("Invalid UTF-8 in location path")?;
+                let metadata = std::fs::metadata(file_path_str).with_context(|| {
+                    "Could not query metadata for the declared file location '{file_path_str}'"
                 })?;
                 if metadata.is_dir() {
                     // If the plugin declared valid file extensions, then we'll filter by those file
@@ -187,7 +195,7 @@ impl<'a> Provider<'a> {
                         .map(|file_type| file_type.extension.as_str())
                         .collect();
 
-                    let walker = WalkDir::new(file_path)
+                    let walker = WalkDir::new(file_path_str)
                         .min_depth(1)
                         .follow_links(true)
                         .same_file_system(false)
@@ -202,25 +210,26 @@ impl<'a> Provider<'a> {
 
                     for candidate in walker {
                         assert!(candidate.path().is_absolute());
-                        let uri = format!("file://{}", candidate.path().to_str().unwrap());
-
-                        // I'm not actually sure if `PathBuf`s always use forward slashes or not,
-                        // but while the original URI does use forward slashes this candidate path
-                        // may not.
-                        #[cfg(windows)]
-                        let uri = uri.replace('\\', "/");
 
                         // TODO: Not quite sure what should be done with errors when crawling
                         //       directories. If the plugin doesn't return an error but also doesn't
                         //       declare any presets then that gets handled gracefully
-                        crawl_uri(uri)?;
+                        crawl(LocationValue::File(
+                            CString::new(
+                                candidate
+                                    .path()
+                                    .to_str()
+                                    .context("Invalid UTF-8 in file path")?,
+                            )
+                            .expect("File path contained null bytes"),
+                        ))?;
                     }
                 } else {
-                    crawl_uri(location.uri.to_uri())?;
+                    crawl(location.value.clone())?;
                 }
             }
-            LocationUri::Internal => {
-                crawl_uri(location.uri.to_uri())?;
+            LocationValue::Internal => {
+                crawl(LocationValue::Internal)?;
             }
         }